fix some more crashes (on OS X only)
git-svn-id: svn://svn.icculus.org/netradiant/trunk@208 61c419a2-8eb2-4b30-bcec-8cead039b335

+++ contrib/brushexport/export.cpp
@@ -149,7 +149,7 @@ void ExportData::GetShaderNameFromShaderPath(const char* path, std::string& name
 		name = tmp.substr(last_slash + 1, tmp.length() - last_slash);
 
 #ifdef _DEBUG
-	globalOutputStream() << "Last: " << last_slash << " " << "length: " << (const unsigned int)tmp.length() << "Name: " << name.c_str() << "\n";
+	globalOutputStream() << "Last: " << (const unsigned int) last_slash << " " << "length: " << (const unsigned int)tmp.length() << "Name: " << name.c_str() << "\n";
 #endif
 }
